Refactor: extract a base class for owning a graph
